requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
opencensus-ext-azure==1.1.12
